requirements.txt
pandas
torch==2.7.1
numpy
transformers
tqdm
scikit-learn
sentencepiece
protobuf
matplotlib
sklearn-crfsuite
xgboost==3.0.0
lightgbm
catboost
openpyxl
tabpfn
tabm
tabicl
shap==0.47.2
optuna
umap-learn
seaborn